+++ packages/isar_generator/lib/src/code_gen/collection_schema_generator.dart
@@ -64,7 +64,6 @@ String generateSchema(ObjectInfo object) {
         {
           'name': index.name,
           'unique': index.unique,
-          'replace': index.replace,
           'properties': [
             for (var indexProperty in index.properties)
               {