+++ packages/isar_generator/lib/src/helper.dart
@@ -87,7 +87,6 @@ List<Index> getIndexAnns(Element element) {
     return Index(
       composite: composite,
       unique: ann.getField('unique')!.toBoolValue()!,
-      replace: ann.getField('replace')!.toBoolValue()!,
       type: indexType,
       caseSensitive: ann.getField('caseSensitive')!.toBoolValue(),
     );