+++ packages/isar_test/lib/bool_model.dart
@@ -1,5 +1,5 @@
+import 'package:flutter/foundation.dart';
 import 'package:isar/isar.dart';
-import 'package:dartx/dartx.dart';
 
 part 'bool_model.g.dart';
 
@@ -11,20 +11,23 @@ class BoolModel {
   @Index()
   bool? field = false;
 
-  @Index()
-  List<bool>? list;
+  @Index(type: IndexType.value)
+  List<bool?>? list;
+
+  @Index(type: IndexType.hash)
+  List<bool?>? hashList;
 
   BoolModel();
 
   @override
   String toString() {
-    return '{field: $field, list: $list}';
+    return '{field: $field, list: $list, hashList: $hashList}';
   }
 
   @override
   bool operator ==(other) {
     return (other as BoolModel).field == field &&
-        list?.length == other.list?.length &&
-        (list?.contentEquals(other.list!) ?? true);
+        listEquals(list, other.list) &&
+        listEquals(hashList, other.hashList);
   }
 }