+++ packages/isar_test/lib/long_model.dart
@@ -1,5 +1,5 @@
+import 'package:flutter/foundation.dart';
 import 'package:isar/isar.dart';
-import 'package:dartx/dartx.dart';
 
 part 'long_model.g.dart';
 
@@ -27,9 +27,7 @@ class LongModel {
   @override
   bool operator ==(other) {
     return (other as LongModel).field == field &&
-        list?.length == other.list?.length &&
-        (list?.contentEquals(other.list!) ?? true) &&
-        hashList?.length == other.hashList?.length &&
-        (hashList?.contentEquals(other.hashList!) ?? true);
+        listEquals(list, other.list) &&
+        listEquals(hashList, other.hashList);
   }
 }