+++ packages/isar_test/lib/string_model.dart
@@ -1,5 +1,5 @@
+import 'package:flutter/foundation.dart';
 import 'package:isar/isar.dart';
-import 'package:dartx/dartx.dart';
 
 part 'string_model.g.dart';
 
@@ -39,12 +39,9 @@ class StringModel {
     if (other is StringModel) {
       return field == other.field &&
           hashField == other.hashField &&
-          list?.length == other.list?.length &&
-          (list?.contentEquals(other.list!) ?? true) &&
-          hashList?.length == other.hashList?.length &&
-          (hashList?.contentEquals(other.hashList!) ?? true) &&
-          hashElementList?.length == other.hashElementList?.length &&
-          (hashElementList?.contentEquals(other.hashElementList!) ?? true);
+          listEquals(list, other.list) &&
+          listEquals(hashList, other.hashList) &&
+          listEquals(hashElementList, other.hashElementList);
     }
     return false;
   }